requirements.txt
spotipy==2.23.0
python-dotenv==1.0.0
pyyaml==6.0.1
requests==2.31.0
pandas
gitpython==3.1.41
